move JoinRequestKeys to own file + unit tests
